Hello World - Video Contribution
In this video, I have discussed a very simple topic in regards with JavaScript, i.e., how to print Hello World, both in Browser's Console, as well as, a Text Editor. This video has been made as part of my contribution to GWOC'21 - GirlScript Winter Of Contributing. It's basically a three - months long open source event, which is led by GirlScript Foundation, which focuses in creating the World's largest multilingual content repository, which will be made available to all for free of cost.