Update Egg_Dropping.md
We wish to know which stories in a no_of_floor -story building are safe to drop eggs from, 
and which will cause the eggs to break on landing. We make a few assumptions:
An egg that survives a fall can be used again. 
A broken egg must be discarded. 
The effect of a fall is the same for all eggs. 
If an egg breaks when dropped, then it would break if dropped from a higher floor. 
If an egg survives a fall then it would survive a shorter fall. 
It is not ruled out that the first-floor windows break eggs, nor is it ruled out that the 36th-floor do not cause an egg to break.
If only one egg is available and we wish to be sure of obtaining the right result, the experiment can be 
carried out in only one way. Drop the egg from the first-floor window; if it survives, drop it from the 
second-floor window. Continue upward until it breaks. 
THE TASK IS TO FIND MINIMUM NUMBER OF TRIALS IN WORST CASE TO FIND WHICH FLOORS ARE SAFE FOR DROPPING EGG

This is one of the best practice problems for recursion! :)
Good luck!