Uploaded Kadane's Algorithm in cpp
The basic idea of Kadane’s algorithm is to find all positive contiguous segments of the array (max_ending_here is used for this). At the same time, keep a track of maximum sum contiguous segment among all positive segments (max_so_far is used for this). Each time we get a positive-sum compare it with max_so_far and update max_so_far if it is greater than max_so_far